refactor(configure): Converted configure cards into components
